content-security-policy/to-port-from-blink/report-blocked-file-uri-expected.txt

CONSOLE ERROR: Refused to load the image 'compass.jpg' because it violates the following Content Security Policy directive: "img-src 'none'".

PingLoader dispatched to 'http://127.0.0.1:8000/security/contentSecurityPolicy/resources/save-report.php?test=report-blocked-file-uri.html'.
CSP report received:
CONTENT_TYPE: application/csp-report
HTTP_REFERER: http://127.0.0.1:8000/security/contentSecurityPolicy/report-blocked-file-uri.html
REQUEST_METHOD: POST
=== POST DATA ===
{"csp-report":{"document-uri":"http://127.0.0.1:8000/security/contentSecurityPolicy/report-blocked-file-uri.html","referrer":"","violated-directive":"img-src 'none'","effective-directive":"img-src","original-policy":"img-src 'none'; report-uri resources/save-report.php?test=report-blocked-file-uri.html","blocked-uri":"file","status-code":200}}
